# Use a single lock for writing to ConcurrentPipeWriter