--- a/src/Servers/Kestrel/Core/src/Internal/Http2/Http2FrameWriter.cs
+++ b/src/Servers/Kestrel/Core/src/Internal/Http2/Http2FrameWriter.cs
@@ -24,6 +24,7 @@ namespace Microsoft.AspNetCore.Server.Kestrel.Core.Internal.Http2
         // Literal Header Field without Indexing - Indexed Name (Index 8 - :status)
         private static ReadOnlySpan<byte> _continueBytes => new byte[] { 0x08, 0x03, (byte)'1', (byte)'0', (byte)'0' };
 
+        private readonly object _writeLock = new object();
         private readonly Http2Frame _outgoingFrame;
         private readonly HPackEncoder _hpackEncoder = new HPackEncoder();
         private readonly ConcurrentPipeWriter _outputWriter;
@@ -55,7 +56,7 @@ namespace Microsoft.AspNetCore.Server.Kestrel.Core.Internal.Http2
             IKestrelTrace log)
         {
             // Allow appending more data to the PipeWriter when a flush is pending.
-            _outputWriter = new ConcurrentPipeWriter(outputPipeWriter, memoryPool);
+            _outputWriter = new ConcurrentPipeWriter(outputPipeWriter, memoryPool, _writeLock);
             _connectionContext = connectionContext;
             _http2Connection = http2Connection;
             _connectionOutputFlowControl = connectionOutputFlowControl;
@@ -70,7 +71,7 @@ namespace Microsoft.AspNetCore.Server.Kestrel.Core.Internal.Http2
 
         public void UpdateMaxFrameSize(uint maxFrameSize)
         {
-            lock (_outputWriter.Sync)
+            lock (_writeLock)
             {
                 if (_maxFrameSize != maxFrameSize)
                 {
@@ -82,7 +83,7 @@ namespace Microsoft.AspNetCore.Server.Kestrel.Core.Internal.Http2
 
         public void Complete()
         {
-            lock (_outputWriter.Sync)
+            lock (_writeLock)
             {
                 if (_completed)
                 {
@@ -97,7 +98,7 @@ namespace Microsoft.AspNetCore.Server.Kestrel.Core.Internal.Http2
 
         public void Abort(ConnectionAbortedException error)
         {
-            lock (_outputWriter.Sync)
+            lock (_writeLock)
             {
                 if (_aborted)
                 {
@@ -113,7 +114,7 @@ namespace Microsoft.AspNetCore.Server.Kestrel.Core.Internal.Http2
 
         public ValueTask<FlushResult> FlushAsync(IHttpOutputAborter outputAborter, CancellationToken cancellationToken)
         {
-            lock (_outputWriter.Sync)
+            lock (_writeLock)
             {
                 if (_completed)
                 {
@@ -129,7 +130,7 @@ namespace Microsoft.AspNetCore.Server.Kestrel.Core.Internal.Http2
 
         public ValueTask<FlushResult> Write100ContinueAsync(int streamId)
         {
-            lock (_outputWriter.Sync)
+            lock (_writeLock)
             {
                 if (_completed)
                 {
@@ -160,7 +161,7 @@ namespace Microsoft.AspNetCore.Server.Kestrel.Core.Internal.Http2
         */
         public void WriteResponseHeaders(int streamId, int statusCode, Http2HeadersFrameFlags headerFrameFlags, IHeaderDictionary headers)
         {
-            lock (_outputWriter.Sync)
+            lock (_writeLock)
             {
                 if (_completed)
                 {
@@ -185,7 +186,7 @@ namespace Microsoft.AspNetCore.Server.Kestrel.Core.Internal.Http2
 
         public ValueTask<FlushResult> WriteResponseTrailers(int streamId, HttpResponseTrailers headers)
         {
-            lock (_outputWriter.Sync)
+            lock (_writeLock)
             {
                 if (_completed)
                 {
@@ -243,7 +244,7 @@ namespace Microsoft.AspNetCore.Server.Kestrel.Core.Internal.Http2
             // The Length property of a ReadOnlySequence can be expensive, so we cache the value.
             var dataLength = data.Length;
 
-            lock (_outputWriter.Sync)
+            lock (_writeLock)
             {
                 if (_completed || flowControl.IsAborted)
                 {
@@ -357,7 +358,7 @@ namespace Microsoft.AspNetCore.Server.Kestrel.Core.Internal.Http2
                 OutputFlowControlAwaitable availabilityAwaitable;
                 var writeTask = default(ValueTask<FlushResult>);
 
-                lock (_outputWriter.Sync)
+                lock (_writeLock)
                 {
                     if (_completed || flowControl.IsAborted)
                     {
@@ -432,7 +433,7 @@ namespace Microsoft.AspNetCore.Server.Kestrel.Core.Internal.Http2
         */
         public ValueTask<FlushResult> WriteWindowUpdateAsync(int streamId, int sizeIncrement)
         {
-            lock (_outputWriter.Sync)
+            lock (_writeLock)
             {
                 if (_completed)
                 {
@@ -455,7 +456,7 @@ namespace Microsoft.AspNetCore.Server.Kestrel.Core.Internal.Http2
         */
         public ValueTask<FlushResult> WriteRstStreamAsync(int streamId, Http2ErrorCode errorCode)
         {
-            lock (_outputWriter.Sync)
+            lock (_writeLock)
             {
                 if (_completed)
                 {
@@ -482,7 +483,7 @@ namespace Microsoft.AspNetCore.Server.Kestrel.Core.Internal.Http2
         */
         public ValueTask<FlushResult> WriteSettingsAsync(IList<Http2PeerSetting> settings)
         {
-            lock (_outputWriter.Sync)
+            lock (_writeLock)
             {
                 if (_completed)
                 {
@@ -515,7 +516,7 @@ namespace Microsoft.AspNetCore.Server.Kestrel.Core.Internal.Http2
         // No payload
         public ValueTask<FlushResult> WriteSettingsAckAsync()
         {
-            lock (_outputWriter.Sync)
+            lock (_writeLock)
             {
                 if (_completed)
                 {
@@ -537,7 +538,7 @@ namespace Microsoft.AspNetCore.Server.Kestrel.Core.Internal.Http2
         */
         public ValueTask<FlushResult> WritePingAsync(Http2PingFrameFlags flags, in ReadOnlySequence<byte> payload)
         {
-            lock (_outputWriter.Sync)
+            lock (_writeLock)
             {
                 if (_completed)
                 {
@@ -567,7 +568,7 @@ namespace Microsoft.AspNetCore.Server.Kestrel.Core.Internal.Http2
         */
         public ValueTask<FlushResult> WriteGoAwayAsync(int lastStreamId, Http2ErrorCode errorCode)
         {
-            lock (_outputWriter.Sync)
+            lock (_writeLock)
             {
                 if (_completed)
                 {
@@ -633,7 +634,7 @@ namespace Microsoft.AspNetCore.Server.Kestrel.Core.Internal.Http2
 
         public bool TryUpdateConnectionWindow(int bytes)
         {
-            lock (_outputWriter.Sync)
+            lock (_writeLock)
             {
                 return _connectionOutputFlowControl.TryUpdateWindow(bytes);
             }
@@ -641,7 +642,7 @@ namespace Microsoft.AspNetCore.Server.Kestrel.Core.Internal.Http2
 
         public bool TryUpdateStreamWindow(StreamOutputFlowControl flowControl, int bytes)
         {
-            lock (_outputWriter.Sync)
+            lock (_writeLock)
             {
                 return flowControl.TryUpdateWindow(bytes);
             }
@@ -649,7 +650,7 @@ namespace Microsoft.AspNetCore.Server.Kestrel.Core.Internal.Http2
 
         public void AbortPendingStreamDataWrites(StreamOutputFlowControl flowControl)
         {
-            lock (_outputWriter.Sync)
+            lock (_writeLock)
             {
                 flowControl.Abort();
             }