@halter73 2019-07-17
I'd appreciate it ff we could wait for the ~2% regression this fixes to show up on aka.ms/aspnet/benchmarks before merging.

> I feel like there is something that slows down first request when passing in output files.

That's why when I run my comparison, I use output files for both master and the branch I'm testing.