@halter73 2019-07-17
Also, are you running with the physical windows server and linux load client? I was getting much closer to 5 million RPS when I benchmarked lock sharing, but my baseline was also much higher.