@jkotalik 2019-07-17
Linux for both.